model_server/requirements.txt
numpy==1.23.0
fastapi==0.68.2
pydantic==1.10.12
uvicorn==0.15.0
torch==2.0.1
transformers==4.35.2
openvino
av
python-dotenv
pillow
